requirements.txt
# AI & Computer Vision
torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128
ultralytics==8.3.155
opencv-python==4.11.0.86
numpy==2.3.0  # ONNXRuntime-GPU 1.22.0互換
Pillow==11.2.1

# ONNX Runtime - GPU版優先（setup_onnx_runtime.pyで自動選択）
# CUDA環境: onnxruntime-gpu==1.22.0 (CUDA 12.8完全対応)
# CPU環境:  onnxruntime==1.22.0
# 
# ⚠️ 重要: pip install実行後、ビルド前に必ず以下を実行してください：
#   cd builds && python setup_onnx_runtime.py
# 
onnxruntime==1.22.0  # setupスクリプトで適切なバージョンに置換される

# Segmentation
segment-anything-py==1.0.1

# NSFW Detection
nudenet==3.4.2

# GUI Framework
tkinter-tooltip==3.1.2  # tkinter is standard library

# Windows Integration
pywin32==310   # for Windows COM integration (desktop shortcuts)

# Utilities
requests==2.32.4
urllib3==2.4.0
tqdm==4.67.1

# Additional dependencies for proper functioning
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.15.3
six==1.17.0
sympy==1.13.3
typing_extensions==4.14.0
tzdata==2025.2
ultralytics-thop==2.0.14

# For exe packaging (development only)
# pyinstaller>=5.10.0  # commented out (not needed for distribution)

pynput>=1.7.6
nudenet>=3.4
python-dotenv>=1.0.0

# Encryption for distribution
cryptography>=41.0.0 